Prevening compiler warning

+++ src/gwin/gwin_widget.c
@@ -318,7 +318,7 @@ static void gwidgetEvent(void *param, GEvent *pe) {
 			return;
 
 		// Use the very simplest possible focus rectangle for now
-		for (i = 0; i < GWIN_FOCUS_HIGHLIGHT_WIDTH; i++) {
+		for (i = 0; i < (uint16_t)GWIN_FOCUS_HIGHLIGHT_WIDTH; i++) {
 			gdispGDrawBox(gx->g.display, gx->g.x+x+i, gx->g.y+y+i, cx-2*i, cy-2*i, gx->pstyle->focus);
 		}
 	}
